feat: conversion from ast to HIR
There is a bunch of changes and programming that I can't be bothered to break into seperate commits so here we are.

The main changes are the moving of the ID types into a seperate crate to prevent circular dependencies and the implemenation of code and traits
that convert from the AST to HIR representations.