feat: add initial trait support in backend
This adds both inital support for trait and impl blocks into the backend i.e. the item tree.
Notably I still have to add parsing for these structures. I was also working on rewriting the
type resolver to use these traits but to do so I plan on creating a "std" library that gets magically
imported and hence haven't included those files in this  commit.

I also changed how the data  around signatures and sources is stored such that data can be more easily gained and
the signatures are more specific.